Factorized the espeak backends, fixed unit tests
